Added Alert for the favorite

+++ components/Favorite.js
@@ -1,6 +1,6 @@
 import React, {Component} from "react";
-import {View, FlatList, Text} from "react-native";
-import {ListItem, Tile} from "react-native-elements";
+import {View, FlatList, Text, Alert} from "react-native";
+import {ListItem} from "react-native-elements";
 import {connect} from 'react-redux'
 import {baseUrl} from "../shared/baseUrl";
 import {Loading} from "./Loading";
@@ -30,7 +30,25 @@ class Favorites extends Component {
                 {
                     text: 'Delete',
                     type: 'delete',
-                    onPress: () => this.props.deleteFavorite(item.id)
+                    onPress: () => {
+                        Alert.alert(
+                            'Delete favorite?',
+                            'Are you sure you wish to delete the favorite dish ' + item.name + '?',
+                            [
+                                {
+                                    text: 'Cancel',
+                                    onPress: () => console.log(item.name + ' not deleted'),
+                                    style: ' cancel'
+                                },
+                                {
+                                    text: 'OK',
+                                    onPress: () => this.props.deleteFavorite(item.id),
+                                    style: ''
+                                }
+                            ],
+                            {cancelable: false}
+                        )
+                    }
                 }
             ]
             return (
